+++ projects/ids-enterprise-ng/src/lib/fileupload-advanced/soho-fileupload-advanced.spec.ts
@@ -1,6 +1,6 @@
 /// <reference path="soho-fileupload-advanced.d.ts" />
 
-import {ComponentFixture, fakeAsync, TestBed, tick} from '@angular/core/testing';
+import { ComponentFixture, fakeAsync, TestBed, tick } from '@angular/core/testing';
 import { By } from '@angular/platform-browser';
 import { Component, DebugElement, ViewChild } from '@angular/core';
 import { FormsModule } from '@angular/forms';
@@ -9,13 +9,14 @@ import { SohoFileUploadAdvancedModule } from './soho-fileupload-advanced.module'
 import { SohoFileUploadAdvancedComponent } from './soho-fileupload-advanced.component';
 
 @Component({
-  template: `<div soho-fileupload-advanced></div>`
+  template: `
+    <div soho-fileupload-advanced></div>`
 })
 class SohoFileUploadAdvancedTestComponent {
   @ViewChild(SohoFileUploadAdvancedComponent) fileuploadadvanced: SohoFileUploadAdvancedComponent;
 }
 
-describe('Soho File Upload Advanced Render', () => {
+fdescribe('Soho File Upload Advanced Render', () => {
   let fileuploadadvanced: SohoFileUploadAdvancedComponent;
   let component: SohoFileUploadAdvancedTestComponent;
   let fixture: ComponentFixture<SohoFileUploadAdvancedTestComponent>;
@@ -34,7 +35,8 @@ describe('Soho File Upload Advanced Render', () => {
     de = fixture.debugElement.query(By.css('div[soho-fileupload-advanced]'));
     el = de.nativeElement;
 
-    fixture.detectChanges();
+    fixture.detectChanges(true);
+
     fileuploadadvanced = component.fileuploadadvanced;
   });
 
@@ -42,32 +44,32 @@ describe('Soho File Upload Advanced Render', () => {
     fixture.detectChanges();
     fileuploadadvanced.disabled = true;
     fixture.detectChanges();
-    // expect(el.hasAttribute('is-disabled')).toBeTruthy();
     fileuploadadvanced.disabled = false;
     fixture.detectChanges();
+
+    expect(fileuploadadvanced.disabled).toBeFalsy();
   });
 
   it('@Input() errorMaxFileSize', fakeAsync(() => {
-     const spy = spyOn(fixture.changeDetectorRef, 'markForCheck');
-
-     const spy2 = spyOn((fileuploadadvanced as any).fileuploadadvanced, 'updated');
+    const markForRefresh = spyOn(fileuploadadvanced, 'markForRefresh').and.callThrough();
+    const updatedSpy = spyOn((fileuploadadvanced as any).fileuploadadvanced, 'updated');
 
     // Check setting error messages - needs to update component.
-    fileuploadadvanced.errorMaxFileSize = 'File too big';
+    fileuploadadvanced.errorMaxFileSize = 'Too many';
 
-    expect(fileuploadadvanced.errorMaxFileSize).toEqual('File too big');
+    expect(fileuploadadvanced.errorMaxFileSize).toEqual('Too many');
 
-    tick();
-    fixture.detectChanges();
-
-    expect(spy).toHaveBeenCalled();
-    expect(spy2).toHaveBeenCalled();
+    fixture.whenStable().then(() => {
+      tick();
+      fixture.detectChanges();
+      expect(markForRefresh).toHaveBeenCalled();
+      expect(updatedSpy).toHaveBeenCalled();
+    });
   }));
 
-  xit('@Input() errorMaxFilesInProcess', fakeAsync(() => {
-    const spy = spyOn(fixture.changeDetectorRef, 'markForCheck');
-
-    const spy2 = spyOn((fileuploadadvanced as any).fileuploadadvanced, 'updated');
+  it('@Input() errorMaxFilesInProcess', fakeAsync(() => {
+    const markForRefresh = spyOn(fileuploadadvanced, 'markForRefresh').and.callThrough();
+    const updatedSpy = spyOn((fileuploadadvanced as any).fileuploadadvanced, 'updated');
 
     // Check setting error messages - needs to update component.
     fileuploadadvanced.errorMaxFilesInProcess = 'Too many';
@@ -77,9 +79,8 @@ describe('Soho File Upload Advanced Render', () => {
     fixture.whenStable().then(() => {
       tick();
       fixture.detectChanges();
-      tick();
-      expect(spy).toHaveBeenCalled();
-        expect(spy2).toHaveBeenCalled();
-      });
+      expect(markForRefresh).toHaveBeenCalled();
+      expect(updatedSpy).toHaveBeenCalled();
+    });
   }));
 });