+++ projects/ids-enterprise-ng/src/lib/tag/soho-tag.component.ts
@@ -2,7 +2,6 @@
 
 import {
   Component,
-  OnInit,
   Input,
   HostBinding,
   ElementRef,